Working on analysis step
Comparison of one index file against multiple search files in a set will
yield the similarity of that file to the entire set, and summing across
all files in the index set will give total similarity.

This is a quadratic step, but is also fast enough that that shouldn't
matter (famous last words).

Todo: write the README