Removed DEBUG_LEVEL declaration from header debug.h
Variables should never be declared in header files. I moved the declaration
from the header file to the main file, and simply defined an extern int in the
header. This will allow other files to see the (global) variable, and even to
modify it, without raising a linker error if there are additional source
files.