Moved definition of DEBUG_LEVEL into commet_job.c.
This is a better place than in the main routines. This appears to be the only
place in the code where the value gets set to a non-zero value, so it seems
like the ideal place for it to live and for it to get initialized to a
zero-value as well.